reorganize gateway structure
