+++ beacon-chain/BUILD.bazel
@@ -8,6 +8,9 @@ go_library(
         "main.go",
         "usage.go",
     ],
+    data = [
+        "//proto/beacon/rpc/v1:swagger",
+    ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
