+++ beacon-chain/gateway/handlers.go
@@ -1,17 +1,16 @@
-package main
+package gateway
 
 import (
-	"fmt"
 	"net/http"
 	"path"
 	"strings"
-
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/connectivity"
 )
 
-// swaggerServer returns swagger specification files located under "/swagger/"
-func swaggerServer(dir string) http.HandlerFunc {
+// Swagger directory for the runtime files provided by bazel data.
+const swaggerDir = "proto/beacon/rpc/v1/"
+
+// SwaggerServer returns swagger specification files located under "/swagger/"
+func SwaggerServer() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		if !strings.HasSuffix(r.URL.Path, ".swagger.json") {
 			log.Debugf("Not Found: %s", r.URL.Path)
@@ -21,21 +20,7 @@ func swaggerServer(dir string) http.HandlerFunc {
 
 		log.Debugf("Serving %s\n", r.URL.Path)
 		p := strings.TrimPrefix(r.URL.Path, "/swagger/")
-		p = path.Join(dir, p)
+		p = path.Join(swaggerDir, p)
 		http.ServeFile(w, r, p)
-	}
-}
-
-// healthzServer returns a simple health handler which returns ok.
-func healthzServer(conn *grpc.ClientConn) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/plain")
-		if s := conn.GetState(); s != connectivity.Ready {
-			http.Error(w, fmt.Sprintf("grpc server is %s", s), http.StatusBadGateway)
-			return
-		}
-		if _, err := fmt.Fprintln(w, "ok"); err != nil {
-			log.WithError(err).Error("failed to respond to healthz")
-		}
 	}
 }
