+++ beacon-chain/gateway/log.go
@@ -1,0 +1,5 @@
+package gateway
+
+import "github.com/sirupsen/logrus"
+
+var log = logrus.WithField("prefix", "gateway")
