+++ beacon-chain/node/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
         "//beacon-chain/attestation:go_default_library",
         "//beacon-chain/blockchain:go_default_library",
         "//beacon-chain/db:go_default_library",
+        "//beacon-chain/gateway:go_default_library",
         "//beacon-chain/operations:go_default_library",
         "//beacon-chain/powchain:go_default_library",
         "//beacon-chain/rpc:go_default_library",
